Regards #590: Avoiding compilation warnings also in examples code

+++ examples/modified_cuda_samples/vectorAdd_nvrtc/vectorAdd_nvrtc.cpp
@@ -93,7 +93,7 @@ int main(void)
 
 	// Verify that the result vector is correct
 	for (int i = 0; i < numElements; ++i) {
-		if (std::fabs(h_A.get()[i] + h_B.get()[i] - h_C.get()[i]) > 1e-5)  {
+		if (std::fabs(h_A.get()[i] + h_B.get()[i] - h_C.get()[i]) > (float) 1e-5)  {
 			std::cerr << "Result verification failed at element " << i << "\n";
 			exit(EXIT_FAILURE);
 		}
